Adding knn only generation script Adding profiling output to test_all

+++ test_all.py
@@ -1,12 +1,15 @@
 import subprocess
 import os
-# python generate_all_samples.py > test.out
+import sys
+import time
+# python test_all.py > test.out
+
 
 algorithms = ['base', 'omp', 'gpu', 'gpud']
 
 print('Dimension, Seed, Sampling, Point_Count, #_Neighbors, Algorithm, Time, Total_Edge_Count, Missed_Edge_Count, Additional_Edge_Count')
-for N in [10000, 100000, 1000000]:
-    for D in range(2, 8):
+for N in [1000000]:
+    for D in range(7, 10):
         for S in range(0, 10):
             for A in ['uniform', 'normal', 'lhs', 'cvt']:
                 if D < 5:
@@ -23,12 +26,24 @@ for N in [10000, 100000, 1000000]:
                 k_filename = 'data/graphs/knn_{}_{}_{}_{}.txt'.format(A, N, D, S)
 
                 # Generate Point Set
-                result = subprocess.run(['python', 'generate_sample_set.py', str(D), str(S), str(N), str(A), p_filename], stdout=subprocess.PIPE, stderr=subprocess.PIPE, check=True)            
+                if not os.path.exists(p_filename) or not os.stat(p_filename).st_size:
+                    sys.stderr.write(p_filename + ' generation: ')
+                    sys.stderr.flush()
+                    start = time.time()
+                    result = subprocess.run(['python', 'generate_sample_set.py', str(D), str(S), str(N), str(A), p_filename], stdout=subprocess.PIPE, stderr=subprocess.PIPE, check=True)
+                    end = time.time()
+                    sys.stderr.write('{} s\n'.format(end-start))
 
                 # Generate Graph to Prune
-                fptr = open(k_filename, 'w')
-                result = subprocess.run(['../knn-generator/binsrc/getNeighborGraph', '-d', str(D), '-b', '1', '-k', str(K), '-i', p_filename], stdout=subprocess.PIPE, stderr=fptr, check=True)
-                fptr.close()
+                if not os.path.exists(k_filename) or not os.stat(k_filename).st_size:
+                    sys.stderr.write(k_filename + ' generation: ')
+                    sys.stderr.flush()
+                    start = time.time()
+                    fptr = open(k_filename, 'w')
+                    result = subprocess.run(['../knn-generator/binsrc/getNeighborGraph', '-d', str(D), '-b', '1', '-k', str(K), '-i', p_filename], stdout=subprocess.PIPE, stderr=fptr, check=True)
+                    fptr.close()
+                    end = time.time()
+                    sys.stderr.write('{} s\n'.format(end-start))
 
                 # Initialize dictionaries for storing the commands for each
                 # algorithm, the execution times, and the files to store the pruned
@@ -47,16 +62,26 @@ for N in [10000, 100000, 1000000]:
                 # Run the algorithms
                 for alg in algorithms:
                     fptr = open(filenames[alg], 'w')
+                    sys.stderr.write(filenames[alg] + ' computation: ')
+                    sys.stderr.flush()
+                    start = time.time()
                     result = subprocess.run(cmds[alg], stdout=fptr, stderr=subprocess.PIPE, check=True)
+                    end = time.time()
+                    sys.stderr.write('{} s\n'.format(end-start))
                     fptr.close()
                     times[alg] = float(result.stderr.decode('utf-8').strip().split('\n')[-1])
 
+                os.remove(p_filename)
+                # os.remove(k_filename)
                 # Count the total number of edges and the differences between
                 # the base algorithm and the others
                 counts = {}
                 base_only = {}
                 other_only = {}
                 for alg,fname in filenames.items():
+                    sys.stderr.write(fname + ' comparison')
+                    sys.stderr.flush()
+                    start = time.time()
                     if os.path.isfile(fname):
                         eset = set()
 
@@ -80,8 +105,11 @@ for N in [10000, 100000, 1000000]:
                             base_only[alg] = 0
                             other_only[alg] = 0
                         else:
-                            base_only[alg] = len(eset_base.difference(eset))
-                            other_only[alg] = len(eset.difference(eset_base))
+                            base_only[alg] = 'na' #len(eset_base.difference(eset))
+                            other_only[alg] = 'na' #len(eset.difference(eset_base))
+                    end = time.time()
+                    sys.stderr.write('{} s\n'.format(end-start))
 
                 for alg in algorithms:
                     print('{}, {}, {}, {}, {}, {}, {}, {}, {}, {}'.format(D, S, A, N, K, alg, times[alg], counts[alg], base_only[alg], other_only[alg]))
+                    sys.stdout.flush()
